Removed Warnings
git-svn-id: http://isdb1.offis.uni-oldenburg.de/repos/odysseus/trunk@19864 aa0bda67-78ed-0310-a382-c5338b0f89ba

Former-commit-id: 29d7401160a592f7a77947bf8d1ab545d07b26b4

+++ server/wrapper/zeromq/wrapper.zeromq/src/zmq/Ctx.java
@@ -282,8 +282,8 @@ public class Ctx {
                 }
     
                 //  In the unused part of the slot array, create a list of empty slots.
-                for (int i = (int) slot_count - 1;
-                      i >= (int) ios + 2; i--) {
+                for (int i = slot_count - 1;
+                      i >= ios + 2; i--) {
                     empty_slots.add (i);
                     slots [i] = null;
                 }
